Converted everything from ts to js
