+++ app/index.jsx
@@ -1,4 +1,4 @@
-import { LoginScreen } from "@/screens/loginScreen";
+import { LoginScreen } from "../screens/loginScreen";
 
 export default function Index() {
   return <LoginScreen />;
