+++ components/ThemedText.jsx
@@ -1,12 +1,5 @@
-import { Text, type TextProps, StyleSheet } from 'react-native';
-
+import { Text, StyleSheet } from 'react-native';
 import { useThemeColor } from '@/hooks/useThemeColor';
-
-export type ThemedTextProps = TextProps & {
-  lightColor?: string;
-  darkColor?: string;
-  type?: 'default' | 'title' | 'defaultSemiBold' | 'subtitle' | 'link';
-};
 
 export function ThemedText({
   style,
@@ -14,7 +7,7 @@ export function ThemedText({
   darkColor,
   type = 'default',
   ...rest
-}: ThemedTextProps) {
+}) {
   const color = useThemeColor({ light: lightColor, dark: darkColor }, 'text');
 
   return (
