+++ components/ThemedView.jsx
@@ -1,0 +1,11 @@
+import { View} from 'react-native';
+
+import { useThemeColor } from '@/hooks/useThemeColor';
+
+
+
+export function ThemedView({ style, lightColor, darkColor, ...otherProps }) {
+  const backgroundColor = useThemeColor({ light: lightColor, dark: darkColor }, 'background');
+
+  return <View style={[{ backgroundColor }, style]} {...otherProps} />;
+}
