+++ package.json
@@ -22,7 +22,7 @@
     "@react-navigation/native": "^6.0.2",
     "expo": "~51.0.8",
     "expo-build-properties": "~0.12.1",
-    "expo-constants": "~16.0.1",
+    "expo-constants": "~16.0.2",
     "expo-dev-client": "~4.0.14",
     "expo-font": "~12.0.5",
     "expo-linking": "~6.3.1",
@@ -44,12 +44,10 @@
   "devDependencies": {
     "@babel/core": "^7.24.0",
     "@types/jest": "^29.5.12",
-    "@types/react": "~18.2.79",
     "@types/react-test-renderer": "^18.0.7",
     "jest": "^29.4.0",
     "jest-expo": "^51.0.0",
-    "react-test-renderer": "18.2.0",
-    "typescript": "~5.3.3"
+    "react-test-renderer": "18.2.0"
   },
   "private": true
 }
