+++ services/firebaseErrorHandling.jsx
@@ -1,19 +1,5 @@
-export type FirebaseErrorCode =
-  | 'auth/invalid-email'
-  | 'auth/user-disabled'
-  | 'auth/user-not-found'
-  | 'auth/wrong-password'
-  | 'auth/email-already-in-use'
-  | 'auth/weak-password'
-  | 'auth/invalid-credential'
-  | 'auth/too-many-requests'; 
-
-type FirebaseErrorMessages = {
-  [key in FirebaseErrorCode]?: string;
-};
-
-// Lets us customize the error message, rather than use default dfirebase ones
-const errorMessages: FirebaseErrorMessages = {
+// Lets us customize the error message, rather than use default Firebase ones
+const errorMessages = {
   'auth/invalid-email': 'The email address is badly formatted.',
   'auth/user-disabled': 'This user has been disabled.',
   'auth/user-not-found': 'No user found with this email address.',
@@ -24,6 +10,6 @@ const errorMessages: FirebaseErrorMessages = {
   'auth/too-many-requests': 'Too many requests. Please try again later.',
 };
 
-export const getFirebaseErrorMessage = (errorCode: FirebaseErrorCode): string => {
+export const getFirebaseErrorMessage = (errorCode) => {
   return errorMessages[errorCode] || 'An unexpected error occurred. Please try again.';
 };
